// src/helpers/type/enums/french-to-english-type.enum.ts
enum FrenchTypeToEnglish {
  NORMAL = 'Normal',
  COMBAT = 'Fighting',
  VOL = 'Flying',
  POISON = 'Poison',
  SOL = 'Ground',
  ROCHE = 'Rock',
  INSECTE = 'Bug',
  SPECTRE = 'Ghost',
  ACIER = 'Steel',
  FEU = 'Fire',
  EAU = 'Water',
  PLANTE = 'Grass',
  ELECTRIK = 'Electric',
  PSY = 'Psychic',
  GLACE = 'Ice',
  DRAGON = 'Dragon',
  TENEBRES = 'Dark',
  FEE = 'Fairy',
  INCONNU = 'Unknown',
}

export default FrenchTypeToEnglish;