// src/helpers/type/enums/english-to-french-type.enum.ts
enum EnglishTypeToFrench {
  NORMAL = 'Normal',
  FIGHTING = 'Combat',
  FLYING = 'Vol',
  POISON = 'Poison',
  GROUND = "Sol",
  ROCK = 'Roche',
  BUG = 'Insecte',
  GHOST = 'Spectre',
  STEEL = 'Acier',
  FIRE = 'Feu',
  WATER = 'Eau',
  GRASS = 'Plante',
  ELECTRIC = 'Électrik',
  PSYCHIC = 'Psy',
  ICE = 'Glace',
  DRAGON = 'Dragon',
  DARK = 'Ténèbres',
  FAIRY = 'Fée',
  UNKNOWN = 'Inconnu',
}

export default EnglishTypeToFrench;